requirements.txt
# KODA Core — Dependencias Python
# Instalar con: pip install -r requirements.txt
# Requisito adicional del sistema: tesseract-ocr
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   macOS:   brew install tesseract tesseract-lang
#   Ubuntu:  sudo apt install tesseract-ocr tesseract-ocr-spa

# Google APIs
google-genai>=1.0.0
google-cloud-texttospeech>=2.16.0
google-cloud-speech>=2.25.0
googlemaps>=4.10.0          # opcional, solo si se usa navegación Maps

# Cámara y visión
opencv-python>=4.9.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Audio
pyaudio>=0.2.14
pygame>=2.5.0

# Utilidades
python-dotenv>=1.0.0
